+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 * Removed trace logging during a successful set in a `SubModelSelectedItem` binding
 * Added one-way-to-source binding
 * Added function-based API for for one-way bindings
+* Added composable monomorphic dispatch wrapping
 
 #### 4.0.0-beta-41
 * Fixed "backwards typing" (#373) and other bugs (like #371) introduced in 4.0.0-beta-40 with more careful use of the Dispatcher (#374)